fix: disconectiong of user
